third_party/blink/web_tests/fast/table/resize-table-row-expected.txt

Regression(99212): table rows get incorrect height after changing some cells' height
https://bugs.webkit.org/show_bug.cgi?id=74303

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(firstRow, '').getPropertyValue('height') is '40px'
PASS getComputedStyle(firstRow, '').getPropertyValue('height') is '20px'
PASS getComputedStyle(firstRow, '').getPropertyValue('height') is '40px'
PASS successfullyParsed is true

TEST COMPLETE
